Minimal re-structure: move AuthContext on top level.
Update import paths.
