+++ src/context/AppContext.jsx
@@ -2,7 +2,7 @@ import { useState, createContext } from 'react'
 
 export const AppContext = createContext(null)
 
-export function AppContextProvider({ children }) {
+export function AppProvider({ children }) {
   const [state, setState] = useState(null)
 
   return (
