+++ src/context/index.js
@@ -1,0 +1,4 @@
+import { AppContext, AppProvider } from './AppContext'
+import { AuthContext, AuthProvider, useAuth } from './AuthContext'
+
+export { AppContext, AppProvider, AuthContext, AuthProvider, useAuth }
