+++ src/features/index.js
@@ -1,4 +1,4 @@
-import { AuthContext, AuthProvider, useAuth } from './auth/AuthContext'
+import { AuthContext, AuthProvider, useAuth } from '../context/AuthContext'
 import AuthPrivateRoute from './auth/AuthPrivateRoute'
 
 export { AuthContext, AuthProvider, useAuth, AuthPrivateRoute }
